Filtro Estados Entregables

+++ apps/sgta-backend/src/main/java/pucp/edu/pe/sgta/model/EntregableXTema.java
@@ -9,6 +9,8 @@ import pucp.edu.pe.sgta.util.EstadoEntrega;
 
 import java.time.OffsetDateTime;
 import java.util.List;
+import java.math.BigDecimal;
+
 
 @Entity
 @Getter
@@ -44,7 +46,7 @@ public class EntregableXTema {
     private EstadoEntrega estado;
 
     @Column(name = "nota_entregable")
-    private Double notaEntregable;
+    private BigDecimal notaEntregable;
 
     @Column(name = "activo", nullable = false)
     private Boolean activo = true;
